Move builtin-array-fromasync.h to toplevel
Putting it in gen/ is inconvenient for downstream users that vendor or
repackage the source files.
